refactor(daemon): move createSnapshotRunner to side-effect-free module
Addresses codex review P1 (session 019e2af8, 2026-05-15).

The A1b commit exported createSnapshotRunner from src/daemon/index.ts.
That file's bottom invokes main().catch(...) unconditionally at import
time, so importing the helper from a unit test would start a real
daemon: acquire the lock file, open the pipe server, possibly call
process.exit(1) on failure, or leave daemon-side resources running.

Fix by moving the runner into a side-effect-free module
src/daemon/snapshotRunner.ts that:

  - Has no top-level side effects.
  - Accepts getBootId as a callback so it does not need access to the
    cachedBootId / getBootIdSync helpers in index.ts.
  - Carries its own minimal log helper so it does not import index.ts.

src/daemon/index.ts now imports createSnapshotRunner from the new
module and passes a getBootId that returns the cached value. Behavior
is unchanged.

Tests updated to import from '../snapshotRunner' (and pass the new
getBootId option). Full suite: 1167 pass. tsc clean.
