+++ src/daemon/index.ts
@@ -9,6 +9,7 @@ import { StateWriter } from './StateWriter';
 import { ProcessMonitor } from './ProcessMonitor';
 import { Watchdog } from './Watchdog';
 import { selectRecoverableSessions } from './recoverySelector';
+import { createSnapshotRunner } from './snapshotRunner';
 import type { DaemonState } from './types';
 import type { DaemonEvent, DaemonCreateSessionParams, DaemonSessionIdParams, DaemonResizeParams } from '../shared/rpc';
 
@@ -859,50 +860,8 @@ function buildState(sessionManager: DaemonSessionManager): DaemonState {
   };
 }
 
-// === Snapshot runner (Phase A — A1b) ===
-//
-// Returns an async function that dumps every live session's RingBuffer to disk
-// and persists sessions.json. Owns a per-runner re-entrancy flag so concurrent
-// invocations (e.g., scheduled tick fires while a previous tick is still
-// flushing) collapse to a single run.
-//
-// Extracted from the inline 30s setInterval body so it can also be invoked
-// once at spawn time, closing the window where no .buf yet exists on disk.
-// A crash within the first 30 s after daemon start would otherwise leave a
-// recovery loop with no buffer file to restore from.
-export function createSnapshotRunner(
-  sessionManager: DaemonSessionManager,
-  stateWriter: StateWriter,
-): () => Promise<void> {
-  let running = false;
-  return async function runSnapshotOnce(): Promise<void> {
-    if (running) return;
-    const managed = sessionManager.listManagedSessions();
-    const live = managed.filter((m) => m.meta.state !== 'dead');
-    if (live.length === 0) return;
-
-    running = true;
-    stateWriter.ensureBufferDir();
-    try {
-      for (const m of live) {
-        const dumpPath = stateWriter.getBufferDumpPath(m.meta.id);
-        try {
-          await m.ringBuffer.dumpToFile(dumpPath);
-        } catch (err) {
-          log('warn', `Snapshot dump failed for ${m.meta.id}:`, err);
-        }
-      }
-      try {
-        const state = buildState(sessionManager);
-        stateWriter.saveImmediate(state);
-      } catch (err) {
-        log('warn', 'Snapshot state save failed:', err);
-      }
-    } finally {
-      running = false;
-    }
-  };
-}
+// Phase A — A1b snapshot runner lives in ./snapshotRunner so the unit tests
+// can import it without triggering main() at the bottom of this file.
 
 // === Graceful shutdown ===
 
@@ -1092,7 +1051,12 @@ async function main(): Promise<void> {
   // Sequential dumps to avoid simultaneous memory peaks from all buffers at once.
   // The runner is also invoked once immediately below (A1b) to close the
   // 30 s window where no .buf exists yet on disk.
-  const runSnapshotOnce = createSnapshotRunner(sessionManager, stateWriter);
+  const runSnapshotOnce = createSnapshotRunner(sessionManager, stateWriter, {
+    getBootId: () => {
+      if (!cachedBootId) cachedBootId = getBootIdSync();
+      return cachedBootId;
+    },
+  });
   const snapshotInterval = setInterval(() => {
     void runSnapshotOnce();
   }, 30_000);
